feat(ui-kit): split the demo into tokens and components pages
Hash-routed (#/tokens, #/components) so deep links and the back button
work without a router dependency. The tokens page grows resolved-value
labels under every swatch — read from the cascade and re-read via a
MutationObserver on data-theme (a theme prop would race the parent
effect that applies the attribute) — plus the foreground pairs, sidebar
and code colors, and a controls-scale section the old single page never
showed.
