+++ packages/ui-kit/demo/shared.tsx
@@ -1,0 +1,34 @@
+import type { CSSProperties, ReactNode } from 'react';
+
+export function Section({ title, children }: { title: string; children: ReactNode }) {
+  return (
+    <section style={{ display: 'grid', gap: 'var(--space-4)' }}>
+      <h2 style={{ margin: 0, fontSize: 20, fontWeight: 600 }}>{title}</h2>
+      {children}
+    </section>
+  );
+}
+
+export function Row({ children, style }: { children: ReactNode; style?: CSSProperties }) {
+  return (
+    <div
+      style={{
+        display: 'flex',
+        flexWrap: 'wrap',
+        alignItems: 'center',
+        gap: 'var(--space-3)',
+        ...style,
+      }}
+    >
+      {children}
+    </div>
+  );
+}
+
+export function DemoIcon() {
+  return (
+    <svg viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="1.5">
+      <path d="M8 2v12M2 8h12" strokeLinecap="round" />
+    </svg>
+  );
+}
